+++ src/main/java/com/westlake/air/pecs/domain/bean/SimpleScanIndex.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @Data
 public class SimpleScanIndex {
 
-    Double rt;
+    Float rt;
 
     Long start;
